Avoid 16-bit arithmethic in GC arenas (503419, r=dmandelin).
UltraBlame original commit: 38586a845abde69d6dcbb86ab54ea7e6202e1fde

+++ js/src/jsgc.h
@@ -284,9 +284,9 @@ typedef struct JSGCChunkInfo JSGCChunkInfo;
 
 struct JSGCArenaList {
     JSGCArenaInfo   *last;          
-    uint16          lastCount;      
+    uint32          lastCount;      
 
-    uint16          thingSize;      
+    uint32          thingSize;      
 
     JSGCThing       *freeList;      
 };
